refactor(cli): extract validate / status / test to commands/ (closes #573)
Completes the cli/main.py split begun in PR (a):

- validate + _group_secret_findings + _run_connection_test
  → drt/cli/commands/validate.py (~190 LOC)
- status + _print_history → drt/cli/commands/status.py (~120 LOC)
- test_syncs + _SyncTestResult + _test_display_name
  → drt/cli/commands/test.py (~130 LOC, registered as `drt test`
  via @app.command(name="test"))

Total moved: ~440 LOC. cli/main.py shrinks 595 → 166 LOC (-72%; of
those, ~100 LOC are actual code — the balance is the comments
catalogueing where each command lives and the back-compat re-export
block).

The remaining surface in main.py matches the #573 wiring target:
_resolve_profile_name (still used by drt/integrations/_runner.py
and tests), version_callback, the @app.callback() main, the _get_*
back-compat shims from PR (a), and the re-exports from PR (a) + (b).

Back-compat: drt.cli.main re-exports the helpers + TypedDict so
legacy `from drt.cli.main import _print_history` / _SyncTestResult /
_run_connection_test / etc. paths keep working — same shim pattern
as PR (a). Verified: 1259 passed, 1 skipped.

Closes #573. The #546 split is now complete: Phase 1 #565, Phase 2a
#572, Phase 2b PR (a) #579, Phase 2b PR (b) [this PR].
